Relocate browser tests
Move the tests to generic package (`browser` instead of `firefox`).
Rename the base class to `BrowsersTest` (from `FirefoxUnitTest`).

Part of #345 - Run functional tests with Chrome
